Learning Python and understanding Google Workspace
I made this file after continuing my course over at Cisco NetAcad and understanding both Tuples and Dictionaries.

I had already been using the Google Quickstart guides to create the other files in my repo,  but didn't entirely comprehend what they were doing. Now I was able to understand how the response body is  interpreted by Python and how the code is interacting with this specifically to pull the information.

This should allow me to further customize this and even obtain things that are usually requested by customers, such as a list of all the Aliases of a given user. This showed me that 'aliases' is another key from the response and can be pulled and printed.